requirements.txt
flask==2.3.3
langchain==0.3.20
langchain-community==0.3.19
langchain-openai==0.3.8
openai==1.61.1
chromadb==0.6.3
onnxruntime
transformers==4.34.0
sentence-transformers==2.2.2
python-dotenv==1.0.1
google-search-results==2.4.2
huggingface_hub>=0.16.4,<1.0
pydantic>=2.7.4,<3.0.0
pypdf
rank_bm25
